Extract block hierarchy to a core skillsaw.blocks package
The lint-tree block hierarchy (ContentBlock, FrontmatteredBlock,
JsonConfigBlock and ~25 subclasses) lived in the leaf rule module
rules/builtin/content_analysis.py, yet rule.py, lint_tree.py,
lint_target.py, and docs/extractor.py all depended on it — several via
lazy imports to dodge the resulting cycle. Core modules were reaching
into a rule package: an inverted layer.

- Move the block hierarchy to a new core skillsaw/blocks/ package, split
  by family (base, content, frontmatter, json_config, coderabbit,
  promptfoo, gather) so no single file is a god module.
  content_analysis.py re-exports every name for backward compatibility
  and keeps its detectors.
- Promote the promptfoo format helpers to a public
  skillsaw/formats/promptfoo.py; the rule package re-exports them under
  their legacy underscore names.
- Move utils.py to core skillsaw/utils.py (re-export shim left at the old
  path). blocks no longer triggers the rules.builtin package __init__,
  which was the source of the import cycle.
- Repoint core modules at the core modules and drop the lazy-import
  workarounds where the cycle is gone.
- Fix a latent shim-binding regression in test_perf_helpers (it patched
  the utils shim, a no-op after the move) and add test_module_layering
  guarding import-order robustness and the re-export contracts.
- Point docs/custom-rules.md at the canonical skillsaw.blocks.

No behavior change: the moved code is AST-identical to the original
except deliberate import-path repoints.
